pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "slesa", email = "slesarenkod@gmail.com"},
]
dependencies = [
    "flask>=2.2.2",
    "psycopg2-binary>=2.9.3",
    "pytest>=7.1.3",
    "requests>=2.28.1",
    "argparse>=1.4.0",
]
requires-python = ">=3.10"
license = {text = "MIT"}
[project.optional-dependencies]

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool]
[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "flake8>=5.0.4",
    "autopep8>=1.7.0",
]
